Implement nested commands
