+++ src/commands/status/masternode.js
@@ -1,0 +1,56 @@
+const {table} = require('table');
+
+const BaseCommand = require('../../oclif/command/BaseCommand');
+
+const MuteOneLineError = require('../../oclif/errors/MuteOneLineError');
+
+const PRESETS = require('../../presets');
+
+class MasternodeStatusCommand extends BaseCommand {
+  /**
+   * @param {Object} args
+   * @param {Object} flags
+   * @param {DockerCompose} dockerCompose
+   * @return {Promise<void>}
+   */
+  async runWithDependencies(
+    {
+      preset,
+    },
+    flags,
+    dockerCompose,
+  ) {
+    const data = [];
+    let output;
+
+    const tableConfig = {
+      //singleLine: true,
+      drawHorizontalLine: (index, size) => {
+        return index === 0 || index === 1 || index === size;
+      }
+    };
+
+    data.push(['Service', 'Status']);
+    data.push(['Version', (await dockerCompose.execCommand(preset, 'core', 'dashd --version')).out.split('\n')[0]])
+    data.push(['Blocks', (await dockerCompose.execCommand(preset, 'core', 'dash-cli getblockcount')).out.trim()]);
+    // ....
+
+    try {
+      output = table(data, tableConfig);
+      console.log(output);
+    } catch (e) {
+      throw new MuteOneLineError(e);
+    }
+  }
+}
+
+MasternodeStatusCommand.description = 'Show Docker status';
+
+MasternodeStatusCommand.args = [{
+  name: 'preset',
+  required: true,
+  description: 'preset to use',
+  options: Object.values(PRESETS),
+}];
+
+module.exports = MasternodeStatusCommand;
